factorized the mem code
