+++ MEAnalysis/python/Analyzer.py
@@ -1,0 +1,40 @@
+from PhysicsTools.HeppyCore.framework.analyzer import Analyzer
+from TTH.MEAnalysis.VHbbTree import lvec
+class FilterAnalyzer(Analyzer):
+    """
+    A generic analyzer that may filter events.
+    Counts events the number of processed and passing events.
+    """
+    def beginLoop(self, setup):
+        super(FilterAnalyzer, self).beginLoop(setup)
+        self.counters.addCounter("processing")
+        self.counters["processing"].register("processed")
+        self.counters["processing"].register("passes")
+
+class EventIDFilterAnalyzer(FilterAnalyzer):
+    """
+    """
+
+    def __init__(self, cfg_ana, cfg_comp, looperName):
+        super(EventIDFilterAnalyzer, self).__init__(cfg_ana, cfg_comp, looperName)
+        self.conf = cfg_ana._conf
+        self.event_whitelist = self.conf.general.get("eventWhitelist", None)
+
+    def beginLoop(self, setup):
+        super(EventIDFilterAnalyzer, self).beginLoop(setup)
+
+    def process(self, event):
+        self.counters["processing"].inc("processed")
+
+        passes = True
+        if not self.event_whitelist is None:
+            passes = False
+            if (event.input.run, event.input.lumi, event.input.evt) in self.event_whitelist:
+                print "IDFilter", (event.input.run, event.input.lumi, event.input.evt)
+                passes = True
+
+        if passes and "eventboundary" in self.conf.general["verbosity"]:
+            print "---", event.input.run, event.input.lumi, event.input.evt
+        if passes:
+            self.counters["processing"].inc("passes")
+        return passes
