+++ MEAnalysis/python/LeptonAnalyzer.py
@@ -1,0 +1,108 @@
+from TTH.MEAnalysis.Analyzer import FilterAnalyzer
+class LeptonAnalyzer(FilterAnalyzer):
+    """
+    Analyzes leptons and applies single-lepton and di-lepton selection.
+
+    Relies on TTH.MEAnalysis.VHbbTree.EventAnalyzer for inputs.
+
+    Configuration:
+    Conf.leptons[channel][cuttype] where channel=mu,ele, cuttype=tight,loose,(+veto)
+    the lepton cuts must specify pt, eta and isolation cuts.
+
+    Returns:
+    event.good_leptons (list of VHbbTree.selLeptons): contains the leptons that pass the SL XOR DL selection.
+        Leptons are ordered by flavour and pt.
+    event.is_sl, is_dl (bool): specifies if the event passes SL or DL selection.
+
+    """
+    def __init__(self, cfg_ana, cfg_comp, looperName):
+        super(LeptonAnalyzer, self).__init__(cfg_ana, cfg_comp, looperName)
+        self.conf = cfg_ana._conf
+
+    def beginLoop(self, setup):
+        super(LeptonAnalyzer, self).beginLoop(setup)
+
+        self.counters["processing"].register("sl")
+        self.counters["processing"].register("dl")
+        self.counters["processing"].register("slanddl")
+
+        self.counters.addCounter("leptons")
+        self.counters["leptons"].register("any")
+        for l in ["mu", "el"]:
+            for a in ["tight", "loose"]:
+                for b in ["", "_veto"]:
+                    lt = l + "_" + a + b
+                    self.counters["leptons"].register(lt)
+
+    def process(self, event):
+        self.counters["processing"].inc("processed")
+        self.counters["leptons"].inc("any", len(event.selLeptons))
+
+        event.mu = filter(
+            lambda x: abs(x.pdgId) == 13,
+            event.selLeptons,
+        )
+        event.el = filter(
+            lambda x: abs(x.pdgId) == 11,
+            event.selLeptons,
+        )
+
+        for a in ["tight", "loose"]:
+            for b in ["", "_veto"]:
+                sumleps = []
+                for l in ["mu", "el"]:
+                    lepcuts = self.conf.leptons[l][a+b]
+                    incoll = getattr(event, l)
+
+                    leps = filter(
+                        lambda x: (
+                            x.pt > lepcuts["pt"]
+                            and abs(x.eta) < lepcuts["eta"]
+                            and abs(getattr(x, self.conf.leptons[l]["isotype"])) < lepcuts["iso"]
+                        ), incoll
+                    )
+
+                    if b == "_veto":
+                        good = getattr(event, "{0}_{1}".format(l, a))
+                        leps = filter(lambda x: x not in good, leps)
+
+                    if a == "tight":
+                        leps = filter(
+                            lambda x: x.tightId,
+                            leps
+                        )
+                    elif a == "loose":
+                        leps = filter(
+                            lambda x: x.looseIdPOG,
+                            leps
+                        )
+                    lep = sorted(leps, key=lambda x: x.pt, reverse=True)
+                    sumleps += leps
+                    lt = l + "_" + a + b
+                    setattr(event, lt, leps)
+                    setattr(event, "n_"+lt, len(leps))
+                    self.counters["leptons"].inc(lt, len(leps))
+
+                setattr(event, "lep_{0}".format(a+b), sumleps)
+                setattr(event, "n_lep_{0}".format(a+b), len(sumleps))
+
+
+        event.is_sl = (event.n_lep_tight == 1 and event.n_lep_tight_veto == 0)
+        event.is_dl = (event.n_lep_loose == 2 and event.n_lep_loose_veto == 0)
+
+        if event.is_sl:
+            self.counters["processing"].inc("sl")
+            event.good_leptons = event.mu_tight + event.el_tight
+        if event.is_dl:
+            self.counters["processing"].inc("dl")
+            event.good_leptons = event.mu_loose + event.el_loose
+
+        passes = event.is_sl or event.is_dl
+        if event.is_sl and event.is_dl:
+            #print "pathological SL && DL event: {0}".format(event)
+            self.counters["processing"].inc("slanddl")
+            passes = False
+
+        if passes:
+            self.counters["processing"].inc("passes")
+        return passes
